requirements.txt
streamlit==1.35.0
plotly
streamlit-pills
snowflake-snowpark-python
snowflake-ml-python